services/gateway/requirements.txt
fastapi
uvicorn
httpx
tenacity
pybreaker
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
prometheus-client
Deprecated
jinja2